rgw_sal_motr: [CORTX-31877] Support Stats for Multipart & Obj OverWrite case
In multipart upload, each part upload increments the total object size by part size
and count by one. When the complete multipart is called, the object count is reset to
consider all parts as single object.
On abort size of each part is subtracted and the object count is decremented by number
of parts as well.

Object overwrite case is also hanlded with the same PR which decrement the size of old
obj and increment the size of new obj in case of unversioned bucket. Obj count will
remain the same.

Signed-off-by: Sumedh A. Kulkarni <sumedh.a.kulkarni@seagate.com>