rgw_sal_motr: [CORTX-31877] Support Stat counting for Multipart & Obj OverWrite case (#286)
* rgw_sal_motr: [CORTX-31877] Support Stats for Multipart & Obj OverWrite case

In multipart upload, each part upload increments the total object size by part size
and count by one. When the complete multipart is called, the object count is reset to
consider all parts as single object.
On abort size of each part is subtracted and the object count is decremented by number
of parts as well.

Object overwrite case is also hanlded with the same PR which decrement the size of old
obj and increment the size of new obj in case of unversioned bucket. Obj count will
remain the same.

* rgw_sal_motr: [CORTX-31877] handle stats on delete marker delete

Delete marker creating does not increment the stats but deletion of
delete marker was causing object count to go down by 1.
Addressed the same by addition of a delete marker check.

Signed-off-by: Sumedh A. Kulkarni <sumedh.a.kulkarni@seagate.com>