rgw_sal_motr: [CORTX-33148] fix init_metadata_cache bug
We are bound to create MotrMetaCache objects as we are directly using
class methods during CRUD operations without checking if metadata caching
is enabled or not.
Earlier I was trying to not create meta cache objects which was breaking
the CRUD, so now reverting it to create those objects again.

NOTE: Ideally we should create MetaCache objects only if caching is enabled.

Signed-off-by: Sumedh A. Kulkarni <sumedh.a.kulkarni@seagate.com>