rgw, rgw_sal: [CORTX-34059] store size_rounded with object metadata
Object's actual size on a disk is unavailable with the object metadata,
so it is calculated at run-time with the help of layout-id and
the size of an object.

For simple objects, the time complexity of the calculation is O(1), but
for multipart objects, the complexity increases to O(number of parts)

Storing the actual size of objects with object metadata will save this time by
avoiding the iterations over each part info.

Signed-off-by: Sumedh Anantrao Kulkarni <sumedh.a.kulkarni@seagate.com>